# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AutoSeguro — backend do agente
  version: "0.1.0"
  description: |
    Contrato congelado na Fase 0. O frontend é construído contra este documento, não
    contra a implementação — é o que permite que a frente de UI e a de backend rodem
    sem se pisarem.

    **Como ler.** São quatro superfícies, uma por tela, e elas se ligam por id:

    | Superfície | Tela | Critério |
    |---|---|---|
    | conversa (WS `/api/chat`) | `/chat` | E1 · C2 |
    | histórico (`/api/conversations`) | `/admin/conversas` | C4 |
    | saúde da `/quote` (`/api/quote-health`) | `/admin/status` | C2 · C4 |
    | fila de handoff (`/api/handoffs`) | `/admin/handoffs` | C3 · C4 |

    De um handoff se salta para a conversa que o gerou; de uma tentativa falha se
    entende o estado do breaker. É a navegação que demonstra rastreabilidade.

    **Duas invariantes atravessam o documento:**

    1. Toda mensagem que apresenta um valor monetário tem `quote_id` não-nulo,
       apontando para uma cotação com `status: ok`. Uma mensagem com preço e
       `quote_id: null` é bug, e a UI a marca como tal.
    2. Todo texto que sai daqui já passou pelo mascaramento de PII. Não existe
       endpoint que devolva a versão crua — inclusive porque a tela precisa mostrar
       o comportamento correto, não uma versão especial para screenshot.

    Autenticação não faz parte do escopo: o backend só é exposto em rede local pelo
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/conversations:
    post:
      tags: [chat]
      summary: Abre uma conversa
      description: Chamado pela tela `/chat` antes de conectar o WebSocket.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel]
              properties:
                channel: { type: string, enum: [console, web] }
                external_ref:
                  type: string
                  description: |
                    Identificador do lead no canal. Nunca um telefone real — no console
                    é o nome da sessão, no web o id do socket. Repositório público.
      responses:
        "201":
          description: criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }

    get:
      tags: [rastreabilidade]
      summary: Lista conversas
      parameters:
        - { name: state, in: query, schema: { $ref: "#/components/schemas/ConversationState" } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: página de conversas
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ConversationSummary" }
                  next_cursor: { type: [string, "null"] }
  /api/conversations/{conversation_id}:
    get:
      tags: [rastreabilidade]
      summary: Detalhe da conversa — mensagens e linha do tempo das cotações
      description: |
        É a resposta ao critério C4 numa chamada só: cada mensagem com id e status, e
        cada cotação com todas as suas tentativas (`attempt`, `http_status`,
        `latency_ms`, `outcome`).

        `messages` vem ordenado por `index`, **nunca por timestamp**.
      parameters:
        - { name: conversation_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: conversa completa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversationDetail" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
  /api/conversations/{conversation_id}/traces:
    get:
      tags: [rastreabilidade]
      summary: O que cada resposta do agente executou — tools, tokens e duração
      description: |
        Granularidade de **chamada**, que `/api/usage` (granularidade de turno) não tem:
        um turno em que o agente qualificou e cotou aparece como uma linha só na tabela
        de custo e como duas execuções distintas aqui, cada uma com argumentos,
        resultado e duração próprios.

        A fonte é `ai.agno_runs`, escrita pelo Agno. Por isso o item **não tem schema
        congelado**: a forma pertence a uma dependência, e prometer aqui um formato que
        ela pode mudar seria um contrato que não controlamos.

        `argumentos` e `resultado` vêm **mascarados**. O Agno grava `tool_args` como o
        modelo os escreveu — com o CEP cru, porque é o que a `/quote` recebe — e esta
        tabela é a única do sistema que não passa pelo mascaramento na escrita.

        `cache_read` é anulável: o Ollama não reporta o campo, e devolver `0` ali seria
        inventar uma medição.

        Lista vazia — nunca 404 — quando a conversa existe e não tem execução
        registrada.
      parameters:
        - { name: conversation_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: os traces da conversa, em ordem de `index`
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [run_id, index, status, tokens_in, tokens_out, duracao_ms, tools]
                      properties:
                        run_id: { type: string }
                        index: { type: integer }
                        status: { type: string }
                        modelo: { type: string }
                        provider: { type: string }
                        tokens_in: { type: integer }
                        tokens_out: { type: integer }
                        cache_read: { type: integer, nullable: true }
                        duracao_ms: { type: number }
                        tools:
                          type: array
                          items:
                            type: object
                            required: [nome, argumentos, resultado, duracao_ms, erro]
                            properties:
                              nome: { type: string }
                              argumentos: { type: object, additionalProperties: true }
                              resultado: { type: string }
                              duracao_ms: { type: number }
                              erro: { type: boolean }
  /api/conversations/{conversation_id}/mensagens:
    post:
      tags: [operacao]
      summary: O operador responde na conversa que assumiu
      description: |
        A outra ponta do handoff: sem isto a fila é uma lista que ninguém atende.

        **Só em conversa `encaminhado`** — enquanto o agente conduz, ele é o único que
        fala pela empresa, e duas vozes no mesmo turno deixam o lead sem saber com quem
        está falando. Fora desse estado, `409`.

        A mensagem é gravada com autor **`operador`**, distinguível no histórico tanto
        do template determinístico (`sistema`) quanto do modelo (`agente`) — para o
        lead os três saem iguais, de propósito; a distinção é auditoria. O texto passa
        pelo mesmo mascaramento de PII na escrita que todo o resto.

        A entrega ao lead é imediata, pelo WebSocket do chat, para quem estiver com a
        página aberta; quem não estiver a recebe no replay do `hello`.
      parameters:
        - { name: conversation_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1 }
      responses:
        "201":
          description: a mensagem gravada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "409":
          description: a conversa ainda está com o agente
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Erro" }
        "422":
          description: texto vazio
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Erro" }
...
  /api/usage:
    get:
      tags: [operacao]
      summary: Custo e uso de tokens
      description: |
        Alimenta o painel de custo do admin, a partir da tabela `turn_usage`.

        Substitui o Langfuse (`docs/DECISOES-FECHADAS.md` §5): a informação fica na
        própria base da aplicação, disponível para quem rodar `docker compose up`, em
        vez de exigir seis containers e ~5,6 GB de imagens para ser conferida.

        Prova duas coisas com número, não com frase: que o prompt caching funciona
        (`cache_read` > 0 a partir do segundo turno) e quanto custa uma conversa.
      parameters:
        - { name: conversation_id, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        "200":
          description: uso agregado e por conversa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usage" }
...
  /api/handoffs:
    get:
      tags: [operacao]
      summary: Fila de handoff
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/HandoffStatus" } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        "200":
          description: fila
          content:
            application/json:
              schema:
                type: object
                required: [items, pendentes]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Handoff" }
                  pendentes:
                    type: integer
                    description: Alimenta o badge visível de qualquer tela.
  /api/handoffs/{handoff_id}:
    patch:
      tags: [operacao]
      summary: Assume ou resolve um handoff
      description: |
        Sem isto a fila não é operável — é só uma lista que cresce. As transições
        válidas são `pendente → assumido → resolvido` e `pendente → resolvido`.
      parameters:
        - { name: handoff_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: "#/components/schemas/HandoffStatus" }
      responses:
        "200":
          description: atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Handoff" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "409":
          description: transição de estado inválida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Erro" }
  /api/chat/{conversation_id}:
    get:
      tags: [chat]
      summary: "WebSocket: o canal do lead"
      description: |
        **Upgrade para WebSocket.** É a implementação `web` da porta `ChannelAdapter`,
        e é onde a degradação da `/quote` fica visível: quando a cotação demora, o
        servidor emite `typing` e o aviso de espera, e o valor chega depois como
        `message` nova — não como um texto que apareceu por mágica no lugar do "…".

        Cliente → servidor: `{"type":"message","text":"..."}`
        Servidor → cliente: `MessageEvent` | `TypingEvent` | `StateEvent`

        Um `MessageEvent` com valor monetário sempre traz `quote_id`.
      parameters:
        - { name: conversation_id, in: path, required: true, schema: { type: string } }
      responses:
        "101": { description: protocolos trocados }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
...
components:
  responses:
    NaoEncontrado:
      description: não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Erro" }
  schemas:
...
    Erro:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
    ConversationState:
      type: string
      enum: [novo, qualificando, cotando, cotado, fechado, encaminhado]
      description: |
        `encaminhado` é alcançável de qualquer estado — é a saída de todos os gatilhos
        de handoff.
    MessageStatus:
      type: string
      enum: [received, pending, sent, failed, discarded]
...
    QuoteJobStatus:
      type: string
      enum: [pending, ok, refused, failed]
    HandoffStatus:
      type: string
      enum: [pendente, assumido, resolvido]
    Conversation:
      type: object
      required: [id, channel, state, criado_em]
      properties:
        id: { type: string }
        channel: { type: string, enum: [console, web] }
        state: { $ref: "#/components/schemas/ConversationState" }
        criado_em: { type: string, format: date-time }
    LeadProfile:
      type: object
      description: |
        Os cinco campos da qualificação. Todos anuláveis porque a conversa é
        incremental; `campos_faltantes` é o que a UI mostra como progresso.
      properties:
        idade: { type: [integer, "null"] }
        veiculo_ano: { type: [integer, "null"] }
        cep:
          type: [string, "null"]
          description: 8 dígitos, sem máscara. Mascarado na exibição.
        data_inicio: { type: [string, "null"], format: date }
        plano_id: { type: [string, "null"], enum: [essencial, completo, premium, null] }
        campos_faltantes:
          type: array
          items: { type: string }
    ConversationSummary:
      allOf:
        - $ref: "#/components/schemas/Conversation"
        - type: object
          required: [total_mensagens, atualizado_em]
          properties:
            total_mensagens: { type: integer }
            atualizado_em: { type: string, format: date-time }
            ultima_mensagem: { type: [string, "null"] }
            handoff_pendente: { type: boolean }
            ultima_cotacao_status:
              oneOf:
                - $ref: "#/components/schemas/QuoteJobStatus"
                - type: "null"
    ConversationDetail:
      allOf:
        - $ref: "#/components/schemas/Conversation"
        - type: object
          required: [perfil, messages, quotes, handoffs]
          properties:
            perfil: { $ref: "#/components/schemas/LeadProfile" }
            messages:
              type: array
              description: Ordenado por `index`. Nunca por timestamp.
              items: { $ref: "#/components/schemas/Message" }
            quotes:
              type: array
              items: { $ref: "#/components/schemas/Quote" }
            handoffs:
              type: array
              items: { $ref: "#/components/schemas/Handoff" }
    Message:
      type: object
      required: [id, index, autor, tipo, conteudo, status, criado_em]
      properties:
        id: { type: string }
        index: { type: integer, minimum: 0 }
        autor: { type: string, enum: [lead, agente, sistema, operador] }
        tipo: { type: string, enum: [text, image, audio, document] }
        conteudo:
          type: string
          description: Sempre a versão **mascarada**. Não existe endpoint com a crua.
        status: { $ref: "#/components/schemas/MessageStatus" }
        quote_id:
          type: [string, "null"]
          description: |
            Não-nulo quando a mensagem apresenta uma cotação. É o guardrail auditável:
            texto com valor monetário e `quote_id` nulo é bug.
        criado_em: { type: string, format: date-time }
...
    Handoff:
      type: object
      required: [id, conversation_id, trigger, reason, status, disparado_por, criado_em]
      properties:
        id: { type: string }
        conversation_id: { type: string }
        trigger:
          type: string
          description: |
            Qual **regra** disparou. É o que torna o critério explícito e defensável:
            a fila mostra o gatilho, não uma frase gerada. O conjunto de valores é
            fechado pela decisão aberta nº1.
        reason: { type: string }
        summary: { type: [string, "null"] }
        disparado_por:
          type: string
          enum: [regra, modelo]
          description: |
            Um gatilho determinístico (breaker aberto, 422 de recusa) e uma decisão do
            modelo produzem o mesmo sinal — e a fila distingue os dois.
        quote_id:
          type: [string, "null"]
          description: Última cotação tentada. É o que o operador vê primeiro ao assumir.
        ultima_cotacao:
          oneOf:
            - $ref: "#/components/schemas/Quote"
            - type: "null"
        status: { $ref: "#/components/schemas/HandoffStatus" }
        criado_em: { type: string, format: date-time }
        assumido_em: { type: [string, "null"], format: date-time }
        resolved_at: { type: [string, "null"], format: date-time }
    Usage:
      type: object
      required: [total, por_provider, por_conversa]
      properties:
        total:
          $ref: "#/components/schemas/UsageAgregado"
        por_provider:
          type: array
          description: Anthropic e Ollama lado a lado.
          items:
            allOf:
              - type: object
                required: [provider]
                properties:
                  provider: { type: string }
              - $ref: "#/components/schemas/UsageAgregado"
        por_conversa:
          type: array
          items:
            allOf:
              - type: object
                required: [conversation_id]
                properties:
                  conversation_id: { type: string }
                  turnos: { type: integer }
              - $ref: "#/components/schemas/UsageAgregado"
...